gd-nasa-api/pyproject.toml
[tool.poetry]
name = "gd-nasa-pipeline"
version = "0.1.0"
description = ""
authors = ["Nacima Yahiatene <nacima.yahiatene@gnomondigital.com>"]
packages = [
    { include = "gd_nasa_api", from = "src" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
sqlalchemy = ">=1.4.28,<2.0"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.25.0"
fastapi = "^0.108.0"
pydantic = "^2.5.3"
pandas = "^2.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_file = "pytest.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

[tool.poetry.scripts]
api = 'gd_nasa_api.main:run'

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
line_length = 79
lines_after_imports = 2
